update before firebase

+++ Assets/MAIN GAME/MAIN/Scripts/StockItem.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using BestHTTP.JSON.LitJson;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
@@ -9,6 +10,7 @@ public class StockItem : MonoBehaviour
     public MainManager MainManager;
     [Header("UI")]
     public Text Name;
+    public Text Score;
     [Header("Info")]
     public StockItemInfo info;
     public void ClickStockItem()
@@ -16,11 +18,38 @@ public class StockItem : MonoBehaviour
         MainManager.Pannel_Detail.info = info;
         MainManager.Pannel_Detail.gameObject.SetActive(true);
     }
+    public void Setup()
+    {
+        Name.text = info.Name;
+        Score.text = info.Score.ToString();
+    }
+    public void ChangeScore(string variablename)
+    {
+        MainManager.Pannel_ChangeInfo.variable_name = variablename;
+        MainManager.Pannel_ChangeInfo.info = info;
+        MainManager.Pannel_ChangeInfo.gameObject.SetActive(true);
+    }
+    
     public void Delete()
     {
-        SaveLoadManager.Instance.DeleteFolder(Name.text, ()=> 
+        MainManager.Pannel_ConfirmDelete.ShowConfirm(() =>
         {
-            DestroyImmediate(gameObject);
+            SaveLoadManager.Instance.DeleteFolder(Name.text, () =>
+            {
+                DestroyImmediate(gameObject);
+            });
         });
+    }
+    public void CopyData()
+    {
+        string jsondata = JsonMapper.ToJson(info);
+        Debug.Log(jsondata);
+
+        TextEditor editor = new TextEditor
+        {
+            text = jsondata
+        };
+        editor.SelectAll();
+        editor.Copy();
     }
 }
